Separation: Detecting closest boid, Directionless Boid Bugfix
[Seperation]
Each boid is able to find the closest boid (other than itself). This can be used to determine which direction to go to avoid colliding with said boid.

[Directionless Boid]
If a boid is forcibly placed in an obstacle, it cannot go in any direction because every direction results in a "crash". This is fine, although when removed from said wall, the boid is now directionless and cannot start up again. This is fixed by having the default direction be north.